+++ tests/test_mixed_simulator_device.py
@@ -445,7 +445,7 @@ class TestApply:
         ):
             simulator_device_1_wire.apply([qml.PauliX(0), qml.BasisState(np.array([0]), wires=[0])])
 
-    def test_qubit_state_vector_not_at_beginning_error(self, simulator_device_1_wire):
+    def test_state_prep_not_at_beginning_error(self, simulator_device_1_wire):
         """Tests that application of StatePrep raises an error if is not
         the first operation."""
 
@@ -474,7 +474,7 @@ class TestStatePreparationErrorsNonAnalytic:
         ):
             simulator_device_1_wire.apply([qml.BasisState(np.array([0]), wires=[0])])
 
-    def test_qubit_state_vector_not_analytic_error(self, simulator_device_1_wire):
+    def test_state_prep_not_analytic_error(self, simulator_device_1_wire):
         """Tests that application of StatePrep raises an error if the device
         is not in analytic mode."""
 
@@ -482,7 +482,7 @@ class TestStatePreparationErrorsNonAnalytic:
 
         with pytest.raises(
             qml.DeviceError,
-            match="The operations StatePrep and QubitStateVector are only supported in analytic mode.",
+            match="The operator StatePrep is only supported in analytic mode.",
         ):
             simulator_device_1_wire.apply([qml.StatePrep(np.array([0, 1]), wires=[0])])